Fix edit state for existing cards
Previously when you opened an existing card it was assumed that you had edited it. This meant that you could save it, and was prompted to save changes when tapping cancel. Now cancel is replaced with done until changes have been made, and save card is disabled until changes are made.

One bug: when editing a draft save card is enabled when one of the sides is edited. The other side could still be blank.