Fix an issue with preview not updating after edit
Reverts a change (7c0b94a) I made that I thought fixed a crash. The crash was actually introduced in changes made in a later commit (9efc62b), and in the final version of that commit those changes had been fixed. So the original fix did nothing. (The crash was caused by taking the attributes of an empty string).

The changes caused the preview to not update after editing a card.