Fix issues with table view insets
There is no default table view cell style with both a subtitle and a right detail. Which is what I want. Previously I used a custom view for all labels, and figured out where to layout the labels. Where they should be changed with iOS 13.

With this commit only the right detail is added through a custom view.