Autocomplete days without tests
One hang-up for new users is opening the app and not having any cards to review. Then leaving the app for a couple of days, opening it again and still having no cards to review. The problem is that the app demands that you do the test every day to count up the day ticker. Even if there are no cards to review.

This commit adds a mechanism for skipping days. Days are automatically skipped if there were no cards to review, up to the first day were there is a test with cards.