Fix iOS 14 issue with slider thumb
Previously the slider thumb was added as a sublayer to a view I guessed to be the actual thumb. Lol. That was not such a good idea.

Anyhow, now the thumb is set as the actual image. Which is the way you are supposed to do it. The image is generated by drawing the layer that was previously added to the assumed thumb view.